Add basic trend record timer
