+++ app/CMakeLists.txt
@@ -40,7 +40,6 @@ zephyr_code_relocate(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/components/*.c LOC
 zephyr_code_relocate(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/data_module.c LOCATION EXTFLASH_TEXT NOCOPY)
 zephyr_code_relocate(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/max32664d_msbl.c LOCATION EXTFLASH_TEXT NOCOPY)
 
-
 zephyr_code_relocate(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/screens/*.c LOCATION EXTFLASH_TEXT NOCOPY)
 #zephyr_code_relocate(FILES ${WEST_TOPDIR}/application/drivers/sensor/max32664dmax32664d_bl.c LOCATION EXTFLASH_TEXT NOCOPY)
 #zephyr_code_relocate(FILES ${WEST_TOPDIR}/application/drivers/sensor/max32664d/max32664d_msbl.c LOCATION EXTFLASH_TEXT NOCOPY)
