+++ app/src/record_module.h
@@ -1,3 +1,2 @@
 #pragma once
 
-void hpi_rec_write_hour(uint32_t filenumber, struct hpi_hr_trend_day_t hr_data);
